Split logic_script_client and Mapbase user messages into their own files
