+++ sp/src/game/shared/mapbase/mapbase_shared.cpp
@@ -618,48 +618,3 @@ BEGIN_DATADESC( CMapbaseManifestEntity )
 
 END_DATADESC()
 #endif
-
-#ifdef CLIENT_DLL
-void __MsgFunc_CallClientScriptFunction( bf_read &msg )
-{
-	char szFunction[64];
-	if (!msg.ReadString( szFunction, sizeof( szFunction ) ))
-	{
-		CGMsg( 0, CON_GROUP_VSCRIPT, "Unable to read function string\n" );
-	}
-
-	int idx = msg.ReadByte();
-	C_BaseEntity *pEntity = CBaseEntity::Instance( idx );
-
-	if (pEntity)
-	{
-		if (pEntity->m_ScriptScope.IsInitialized())
-		{
-			//CGMsg( 0, CON_GROUP_VSCRIPT, "%s calling function \"%s\"\n", pEntity->GetDebugName(), szFunction );
-			pEntity->CallScriptFunction( szFunction, NULL );
-		}
-		else
-		{
-			CGMsg( 0, CON_GROUP_VSCRIPT, "%s scope not initialized\n", pEntity->GetDebugName() );
-		}
-	}
-	else
-	{
-		CGMsg( 0, CON_GROUP_VSCRIPT, "Clientside entity not found for script function (index %i)\n", idx );
-	}
-}
-
-void HookMapbaseUserMessages( void )
-{
-	HOOK_MESSAGE( CallClientScriptFunction );
-}
-#endif
-
-void RegisterMapbaseUserMessages( void )
-{
-	usermessages->Register( "CallClientScriptFunction", -1 );
-
-#ifdef CLIENT_DLL
-	HookMapbaseUserMessages();
-#endif
-}
